+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * `(/ 0.0 0.0)` return NaN
 * bookmarklet always fetch latest version
 * `case-lambda` hygienic macro (code from R7RS)
+* new `boolean=?` and `port?` functions (R7RS)
 * rational `numerator` and `denominator` function (R5RS)
 * complex `imag-part`, `real-part`, `make-polar`, `angle` and `magnitude` functions (R5RS)
 * new function: `radian->degree` and `degree->radian`